Move component classes to seperate folder
